src/frontend/requirements.txt
gradio==4.44.0
tensorflow==2.17.0
keras==3.6.0
numpy==1.24.3
Pillow==10.0.1
opencv-python-headless==4.8.1.78